requirements.txt
aioprocessing==2.0.1
aioquic==0.9.25
amqp==5.3.1
ansicolors==1.1.8
asgiref==3.10.0
asttokens==2.2.1
beautifulsoup4==4.12.2
bidict==0.23.1
billiard==4.2.2
bleach==6.2.0
boltons==25.0.0
Brotli==1.1.0
bs4==0.0.1
cached-property==1.5.2
celery==5.5.3
certifi==2023.5.7
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.1.0
click==8.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
codex==1.8.17
colorama==0.4.6
comm==0.1.3
confuse==2.1.0
cryptography==46.0.3
cssselect2==0.8.0
Cython==3.1.4
dateparser==1.2.2
debugpy==1.6.7
decorator==5.1.1
decouple==0.0.7
deepdiff==8.6.1
diff-match-patch==20241021
dj-database-url==2.2.0
Django==5.2.8
django-cachalot==2.8.0
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-environ==0.12.0
django-import-export==4.3.14
django-jazzmin==2.6.2
django-rest-registration==0.9.0
django-vite==3.1.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
drf-spectacular==0.28.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.13.1
filetype==1.2.0
fnvhash==0.2.1
future==0.18.3
git-filter-repo==2.47.0
glom==24.11.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
humanfriendly==10.0
humanize==4.14.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.3.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jsonschema==4.25.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kombu==5.5.4
loguru==0.7.3
lxml==6.0.0
markdown-it-py==4.0.0
marshmallow==4.0.1
marshmallow-union==0.1.15.post1
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==2.2.2
openpyxl==3.1.2
orderly-set==5.5.0
packaging==24.2
pandas==2.3.3
parso==0.8.3
pillow==12.0.0
platformdirs==4.1.0
pluggy==1.6.0
priority==2.0.0
prompt-toolkit==3.0.38
protobuf==6.31.0
psutil==7.1.2
psycopg2-binary==2.9.11
pure-eval==0.2.2
PyJWT==2.10.1
pymongo==3.13.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPDF2==1.26.0
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.0.2
rarfile==4.2
redis==3.5.3
referencing==0.37.0
reportlab==4.4.4
requests==2.31.0
requests-cache==0.5.2
rich==14.2.0
rich-argparse==1.7.1
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
service-identity==24.2.0
simplejson==3.20.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.5.3
stack-data==0.6.2
stripe==13.0.1
svglib==1.6.0
tablib==3.9.0
text-unidecode==1.3
texttable==1.7.0
tinycss2==1.4.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.0.2
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.26.6
watchdog==6.0.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.9.0
websockets==11.0.3
whitenoise==6.9.0
xmltodict==1.0.2
zstandard==0.25.0

pydantic>=2.0
